Matrix struct + init biases

+++ include/neurone.h
@@ -5,12 +5,12 @@
 
 typedef struct Neuron Neuron;
 struct Neuron {
-    double *bias;
-    double **weights_in;
-    double **weights_out;
+    float *bias;
+    float **weights_in;
+    float **weights_out;
 };
 
 typedef struct Layer Layer;
-Neuron neuron_new(double *biais, Layer *layer, size_t index);
+Neuron neuron_new(float *biais, Layer *layer, size_t index);
 
 #endif
